refactor(playground): rename lang directories
