+++ app/src/components/templateEditor/language/fString/fStringTemplating.syntax.grammar.d.ts
