+++ app/src/pages/playground/PlaygroundChatTemplate.tsx
@@ -247,7 +247,7 @@ function SortableMessageItem({
             height="100%"
             value={message.content}
             aria-label="Message content"
-            templateLanguage={TemplateLanguages.FString}
+            templateLanguage={TemplateLanguages.Mustache}
             onChange={(val) => {
               updateInstance({
                 instanceId: playgroundInstanceId,
